# pkg/policy/api: Optimize Decision MarshalJSON()