Use strings.Builder instead of fmt.Sprintf() and preallocate the size of
the string so that Go doesn't need to over-allocate if the string ends
up longer than what the buffer growth algorithm predicts.

Result:
```
$ go test -benchmem -run=^$ -bench ^BenchmarkDecisionMarshalJSON$ ./pkg/policy/api > BenchmarkDecisionMarshalJSON_old.txt
$ go test -benchmem -run=^$ -bench ^BenchmarkDecisionMarshalJSON$ ./pkg/policy/api > BenchmarkDecisionMarshalJSON_new.txt
$ benchcmp BenchmarkDecisionMarshalJSON_old.txt BenchmarkDecisionMarshalJSON_new.txt
benchmark                          old ns/op     new ns/op     delta
BenchmarkDecisionMarshalJSON-4     751           157           -79.13%

benchmark                          old allocs     new allocs     delta
BenchmarkDecisionMarshalJSON-4     3              3              +0.00%

benchmark                          old bytes     new bytes     delta
BenchmarkDecisionMarshalJSON-4     40            40            +0.00%
```

issue reference : https://github.com/cilium/cilium/issues/19571

Signed-off-by: MikeLing <sabergeass@gmail.com>