@kaworu 2022-05-09
@MikeLing could you rebase on top of `master` now that https://github.com/cilium/cilium/pull/19716 has been merged please (should fix the CI issues — at least ci-awscni and ci-eks)?